requirements.txt
Django~=5.0.7
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
djangorestframework-simplejwt[crypto]
psycopg2-binary~=2.9.9

